# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NutriMyWay API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Member {
  id: number;
  name: string;
  /** @nullable */
  date_of_joining?: string | null;
  /** @nullable */
  height_cm?: number | null;
  /** @nullable */
  daily_kcal?: number | null;
}

export interface Center {
  id: string;
  name: string;
}

export interface HealthRecord {
  id: number;
  member_id: number;
  /** @nullable */
  center_id?: string | null;
  recorded_at: string;
  /** @nullable */
  weight_kg?: number | null;
  /** @nullable */
  body_fat_pct?: number | null;
  /** @nullable */
  visceral_fat?: number | null;
  /** @nullable */
  bmr?: number | null;
  /** @nullable */
  bmi?: number | null;
  /** @nullable */
  metabolic_age?: number | null;
  /** @nullable */
  muscle_mass_kg?: number | null;
  /** @nullable */
  resting_hr?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface HealthRecordInput {
  /**
     * ISO date or date-time string; defaults to now if omitted
     * @nullable
     */
  recorded_at?: string | null;
  /** @nullable */
  center_id?: string | null;
  /** @nullable */
  weight_kg?: number | null;
  /** @nullable */
  body_fat_pct?: number | null;
  /** @nullable */
  visceral_fat?: number | null;
  /** @nullable */
  bmr?: number | null;
  /** @nullable */
  bmi?: number | null;
  /** @nullable */
  metabolic_age?: number | null;
  /** @nullable */
  muscle_mass_kg?: number | null;
  /** @nullable */
  resting_hr?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface ConsumptionLog {
  id: number;
  member_id: number;
  logged_at?: string;
  meal_slot: string;
  food_item: string;
  /** @nullable */
  quantity_g?: number | null;
  /** @nullable */
  calories_kcal?: number | null;
  /** @nullable */
  protein_g?: number | null;
  /** @nullable */
  carbs_g?: number | null;
  /** @nullable */
  fat_g?: number | null;
  /** @nullable */
  photo_url?: string | null;
  /** @nullable */
  photo_uploaded_at?: string | null;
}

export interface ConsumptionLogInput {
  meal_slot: string;
  food_item: string;
  /** @nullable */
  quantity_g?: number | null;
  /** @nullable */
  calories_kcal?: number | null;
  /** @nullable */
  protein_g?: number | null;
  /** @nullable */
  carbs_g?: number | null;
  /** @nullable */
  fat_g?: number | null;
  /** @nullable */
  photo_url?: string | null;
}

export type DailySummaryLogsBySlot = {[key: string]: ConsumptionLog[]};

export interface DailySummary {
  date: string;
  total_calories: number;
  total_protein: number;
  total_carbs: number;
  total_fat: number;
  target_calories?: number;
  logs_by_slot?: DailySummaryLogsBySlot;
}

export interface BomItem {
  id: number;
  /** @nullable */
  plan_name?: string | null;
  food_item: string;
  /** @nullable */
  quantity_g?: number | null;
  /** @nullable */
  calories_kcal?: number | null;
  /** @nullable */
  protein_g?: number | null;
  /** @nullable */
  carbs_g?: number | null;
  /** @nullable */
  fat_g?: number | null;
}

export interface PackSize {
  id: number;
  item_name: string;
  /** @nullable */
  pack_label?: string | null;
  /** @nullable */
  weight_g?: number | null;
  /** @nullable */
  calories_kcal?: number | null;
}

export interface Issuance {
  id: number;
  member_id: number;
  /** @nullable */
  center_id?: string | null;
  /** @nullable */
  pack_label?: string | null;
  issued_at: string;
  /** @nullable */
  status?: string | null;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. image/jpeg).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type RequestUploadUrl400 = {
  error?: string;
};

export type GetPublicObject404 = {
  error?: string;
};

export type GetConsumptionLogsParams = {
/**
 * Date filter in YYYY-MM-DD format
 */
date?: string;
};

export type GetDailySummaryParams = {
/**
 * Date in YYYY-MM-DD format
 */
date?: string;
};

export type GetBomItemsParams = {
plan?: string;
};