# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NutriMyWay API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members/{memberId}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member profile
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{memberId}/centers:
    get:
      operationId: getMemberCenters
      tags: [members]
      summary: Get centers mapped to member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of centers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Center"
...
  /members/{memberId}/consumption:
    get:
      operationId: getConsumptionLogs
      tags: [members]
      summary: Get consumption logs
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: false
          schema:
            type: string
            description: Date filter in YYYY-MM-DD format
      responses:
        "200":
          description: List of consumption logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsumptionLog"
    post:
      operationId: createConsumptionLog
      tags: [members]
      summary: Add consumption log entry
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumptionLogInput"
      responses:
        "201":
          description: Consumption log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionLog"
  /members/{memberId}/summary:
    get:
      operationId: getDailySummary
      tags: [members]
      summary: Get daily calorie and macro totals
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: false
          schema:
            type: string
            description: Date in YYYY-MM-DD format
      responses:
        "200":
          description: Daily summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"
  /members/{memberId}/issuances:
    get:
      operationId: getMemberIssuances
      tags: [members]
      summary: Get pack issuance history for member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of issuances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issuance"
...
components:
  schemas:
...
    Member:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        date_of_joining:
          type: ["string", "null"]
        height_cm:
          type: ["integer", "null"]
        daily_kcal:
          type: ["integer", "null"]
      required:
        - id
        - name
    Center:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
...
    ConsumptionLog:
      type: object
      properties:
        id:
          type: integer
        member_id:
          type: integer
        logged_at:
          type: string
        meal_slot:
          type: string
        food_item:
          type: string
        quantity_g:
          type: ["number", "null"]
        calories_kcal:
          type: ["number", "null"]
        protein_g:
          type: ["number", "null"]
        carbs_g:
          type: ["number", "null"]
        fat_g:
          type: ["number", "null"]
        photo_url:
          type: ["string", "null"]
        photo_uploaded_at:
          type: ["string", "null"]
      required:
        - id
        - member_id
        - meal_slot
        - food_item
    ConsumptionLogInput:
      type: object
      properties:
        meal_slot:
          type: string
        food_item:
          type: string
        quantity_g:
          type: ["number", "null"]
        calories_kcal:
          type: ["number", "null"]
        protein_g:
          type: ["number", "null"]
        carbs_g:
          type: ["number", "null"]
        fat_g:
          type: ["number", "null"]
        photo_url:
          type: ["string", "null"]
      required:
        - meal_slot
        - food_item
    DailySummary:
      type: object
      properties:
        date:
          type: string
        total_calories:
          type: number
        total_protein:
          type: number
        total_carbs:
          type: number
        total_fat:
          type: number
        target_calories:
          type: number
        logs_by_slot:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/ConsumptionLog"
      required:
        - date
        - total_calories
        - total_protein
        - total_carbs
        - total_fat
...
    Issuance:
      type: object
      properties:
        id:
          type: integer
        member_id:
          type: integer
        center_id:
          type: ["string", "null"]
        pack_label:
          type: ["string", "null"]
        issued_at:
          type: string
        status:
          type: ["string", "null"]
      required:
        - id
        - member_id
        - issued_at
...